@@ -9,8 +9,8 @@
     "lint": "eslint --ext .js --ignore-path .gitignore ."
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@vitejs/plugin-react": "^1.0.7",
