@@ -1,14 +1,14 @@
-/* eslint-disable react/react-in-jsx-scope */
-import "./Button.css";
+import React from "react";
+import "../../styles/Button.css";
 
 // eslint-disable-next-line react/prop-types
-const Button = ({ advice }) => {
+const Button = ({ action }) => {
   return (
     <div>
       <div>
-        <button className="button" onClick={advice}>
-          <svg width="24" height="24" xmlns="http://www.w3.org/2000/svg">
-            <path
+        <button className="button" onClick={action}>
+          <svg className="icon" xmlns="http://www.w3.org/2000/svg">
+            <path className="icon-shape"
               d="M20 0H4a4.005 4.005 0 0 0-4 4v16a4.005 4.005 0 0 0 4 4h16a4.005 4.005 0 0 0 4-4V4a4.005 4.005 0 0 0-4-4ZM7.5 18a1.5 1.5 0 1 1 0-3 1.5 1.5 0 0 1 0 3Zm0-9a1.5 1.5 0 1 1 0-3 1.5 1.5 0 0 1 0 3Zm4.5 4.5a1.5 1.5 0 1 1 0-3 1.5 1.5 0 0 1 0 3Zm4.5 4.5a1.5 1.5 0 1 1 0-3 1.5 1.5 0 0 1 0 3Zm0-9a1.5 1.5 0 1 1 0-3 1.5 1.5 0 0 1 0 3Z"
               fill="#202733"
             />
