@@ -1,4 +1,5 @@
-import './Card.css'
+import React from 'react';
+import '../../styles/Card.css'
 
 const Card = (props) => {
   const classes = 'card ' + props.className;
